feat: Add environment variable for debug filenames
This commit adds a new file `env.rs` where we define the environment variables that MontyC uses. Specifically, it introduces the constant `MONTYC_DEBUG_FILE` which can be set to a truthy value to enable debug filenames in the logging output.

In `main.rs`, we also import the newly created `env` module and update the code to use `env::MONTYC_DEBUG_FILE` instead of directly referencing `"MONTYC_DEBUG_FILE"` when checking if debug filenames should be displayed in the logging output.